requirements.txt
#!/usr/bin/env python3
# Basic Agent Dependencies

# Core AI Agent Framework
crewai>=0.30.11
crewai[tools]

# AWS Bedrock Integration
boto3>=1.34.0
botocore>=1.34.0

# Web Framework
Flask>=2.3.3
Flask-CORS>=4.0.0

# Data Processing
pydantic>=2.0.0

# Development and Testing
python-dotenv>=1.0.0

# Optional: For enhanced logging and monitoring
colorlog>=6.7.0

# Note: Additional dependencies will be added in later phases:
# Phase 2: Real tool integrations (requests, httpx, etc.)
# Phase 3: Investment tools (pandas, numpy, etc.)  
# Phase 4: Production (gunicorn, prometheus-client, etc.) 